Change structure of chat state in redis and how chat_id and timestamp are stored
chat id is partition key and timestamp sort key in the DDB table. When a new chat is opened (new ws connection opened and when new chat is triggered) chat_id and timestamp are generated. This model is stored in the Redis state and saved in the DB when connction is closed and everytime the assistant finishes its response.